api/requirements.txt
# Flask API requirements for Yumi Sugoi Discord Bot Dashboard
# Core Flask components
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-socketio==5.3.6

# Database
SQLAlchemy==2.0.25
alembic==1.13.1

# Redis for real-time communication
redis==5.0.1

# Authentication and security
PyJWT==2.8.0
werkzeug==3.0.1

# System monitoring
psutil==5.9.6

# HTTP client
requests==2.31.0

# WebSocket support
python-socketio==5.10.0
eventlet==0.33.3

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
